backend/requirements.txt
# Supply Chain Orchestrator AI — Python Dependencies
# Python >= 3.11 required

# ── Core Framework ─────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
httpx>=0.28.0

# ── LangChain + LangGraph ─────────────────────────────────────────────────
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0
langgraph-checkpoint>=2.0.0

# ── LLM / AI ──────────────────────────────────────────────────────────────
openai>=1.50.0
tiktoken>=0.8.0
langchain-ollama>=0.1.0
langchain-groq>=0.1.0

# ── External Search ───────────────────────────────────────────────────────
tavily-python>=0.5.0

# ── Vector DB ─────────────────────────────────────────────────────────────
chromadb>=0.5.0
sentence-transformers>=3.0.0

# ── Data Validation ───────────────────────────────────────────────────────
pydantic>=2.9.0
pydantic-settings>=2.5.0

# ── Utilities ─────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
